usaugment-experiments/requirements.txt
albumentations==1.4.18
comet-ml==3.45.0
hydra-core==1.3.2
h5py==3.11.0
imageio[ffmpeg]==0.5.1
ipykernel
ipympl
lightning==2.2.4
monai==1.3.0
matplotlib
nibabel==5.2.1
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl
optuna==4.0.0
pandas==2.2.1
pre-commit
python-dotenv==1.0.1
PyYAML
scikit-image==0.24.0
scikit-learn==1.5.0
scikit-posthocs
simpleitk==2.3.1
seaborn
tabulate
torch==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0